MNIST data proof of concept
Added code to run the Network with the MNIST dataset. The dataset is not included in this project, and must be obtained externally. This is mostly a proof of concept, showing the backpropagation algorithm does work.

Also added ability to have differently sized input and output images with processing grayscale to color images.

Also made some minor documentation changes